requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.104.1
h11==0.14.0
idna==3.4
motor==3.3.1
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-decouple==3.8
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
